Phase-IV-local-Kubernets/backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.14
pydantic==2.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
uvicorn[standard]==0.25.0
gunicorn==21.2.0
openai==1.10.0
bcrypt==4.1.2
pyjwt==2.8.0
python-jose==3.3.0
python-multipart==0.0.6
